Remove the need to tangle the config file on change.
Before, when you changed the config you needed to tangle the file so
the init.el file would be created. Now, on startup, if the config.org
file changed, the file will be tangle.

It also means that the init.el is not needed as it is created the
first time emacs is launched.